refactor: move layouts para pasta layouts
